requirements.txt
torch>=1.12.0
numpy
scikit-learn
transformers
datasets
librosa
soundfile
open_clip_torch
laion-clap
matplotlib
tqdm
psutil